outsource of filters in SurfSpotFilter component
