+++ frontend/src/pages/FavouritesPage.js
@@ -1,18 +1,33 @@
+import styled from "styled-components/macro";
 import SpotList from "../component/SpotList";
 import useFavourites from "../hooks/useFavourites";
+import useFilter from "../hooks/useFilter";
+import SurfSpotFilter from "../component/SurfSpotFilter";
+import React from "react";
 
 export default function FavouritesPage(){
 
     const {favourites, setFavourites} = useFavourites();
-
+    const {handleChangeCountry,
+        handleChangeContinent,
+        continents,
+        countries,
+        filterCountry,
+        filterContinent} = useFilter(favourites);
 
     return(
-        <div>
+        <Wrapper>
             <div>
-               <p>Hallo, bin ich noch wach?</p>
+                <h3>Your Favourites</h3>
             </div>
             <div>
-               <p>Hallo, bin ich noch wach?</p>
+                <SurfSpotFilter filterContinent={filterContinent}
+                                filterCountry={filterCountry}
+                                handleChangeContinent={handleChangeContinent}
+                                handleChangeCountry={handleChangeCountry}
+                                continents={continents}
+                                countries={countries}
+                />
             </div>
             <div>
                <SpotList surfSpots={favourites}
@@ -21,7 +36,8 @@ export default function FavouritesPage(){
                          filterCountry={{name: ''}}
                          filterContinent={{name: ''}}/>
             </div>
-        </div>
+        </Wrapper>
     )
-
 }
+const Wrapper = styled.div`
+`
